api/requirements.txt
Flask
flask_hal
gunicorn
python-dotenv
redis
bcrypt
pyjwt